[l1preproc] code adaptions to get l1p preprocessor to work with python 3 pysiral version

+++ pysiral/retracker.py
@@ -7,9 +7,8 @@ Created on Fri Jul 31 15:48:58 2015
 
 # cythonized bottleneck functions for cTFMRA
 try:
-    from pysiral.bnfunc.cytfmra import (cytfmra_findpeaks, cytfmra_interpolate,
-                                        cytfmra_wfm_noise_level,
-                                        cytfmra_normalize_wfm)
+    from .bnfunc.cytfmra import (cytfmra_findpeaks, cytfmra_interpolate,
+                                 cytfmra_wfm_noise_level, cytfmra_normalize_wfm)
     CYTFMRA_OK = True
 except:
     CYTFMRA_OK = False
@@ -1207,7 +1206,7 @@ def smooth(x, window):
 
 def bnsmooth(x, window):
     """ Bottleneck implementation of the IDL SMOOTH function """
-    pad = (window-1)/2
+    pad = int((window-1)/2)
     n = len(x)
     xpad = np.ndarray(shape=(n+window))
     xpad[0:pad] = 0.0
